\name{BSphe1}
\alias{BSphe1}
\docType{data}
\title{Simulated phenotypic data}
\description{
Phenotypic data simulated with BSdat which has one observation per genotype
}
\usage{data(BSphe1)}
\format{
A data frame with three columns:
\describe{
\item{ID:}{the ID for each individual. There is a single replicated for each of 250 individuals}
\item{phenotype:}{the phenotype value. It is generated from the same
QTL effects as in BSdat}
}
}
\references{
Broman, KW and Speed TP. 2002. A model selection approach for the identification of quantitative trait loci in experimental crosses. JRSS-B 64:641-656.
}
\keyword{datasets}
